Construct AWS S3 Call

Constructs GET information string for AWS S3 Bucket

make_aws_call(
  path_to_file = "/",
  bucket = "hcp-openaccess",
  region = "us-east-1",
  access_key = NULL,
  secret_key = NULL,
  lifetime_minutes = 5,
  query = NULL,
  verb = "GET",
  sign = TRUE
)

Arguments

path_to_file
Path to file on HCP S3 Bucket

bucket
Bucket to download from

region
Region of S3 Bucket

access_key
Amazon S3 Access Key

secret_key
Amazon S3 Secret Key

lifetime_minutes
Time that connection can be opened

query
additional query to add to verb command

verb
httr VERB to be used

sign
Should the url be signed?

Value

Character of the url to be passed to httr `VERB`s

Examples

if (have_aws_key()){
path_to_file <- paste0(
   "HCP_900/100206/MNINonLinear/",
   "100206.164k_fs_LR.wb.spec")
result = make_aws_call(path_to_file)
# don't want this stuff printed because it has keys
result$headers$secret_key = NULL
result$headers$access_key = NULL
result$query$AWSAccessKeyId = NULL
result$query$Signature = NULL
result
}

#> $url
#> [1] "https://hcp-openaccess.s3.us-east-1.amazonaws.com/"
#> 
#> $path
#> [1] "HCP_900/100206/MNINonLinear/100206.164k_fs_LR.wb.spec"
#> 
#> $query
#> $query$Expires
#> [1] 1603111733
#> 
#> 
#> $headers
#> $headers$default_region
#> [1] "us-east-1"
#> 
#> 
#> attr(,"canonical_string")
#> [1] "GET\n\n\n1603111733\n/HCP_900/100206/MNINonLinear/100206.164k_fs_LR.wb.spec"